renumbering made correct
