--- tex_files/make_all
@@ -1,16 +1,0 @@
-# source ~/venv/bin/activate
-
-pdflatex tutorial_a4
-pythontex tutorial_a4
-pdflatex tutorial_a4
-
-pdflatex tutorial_mobile
-pythontex tutorial_mobile
-pdflatex tutorial_mobile
-
-rm hint.tex ans.tex
-rm x.log
-
-mv tutorial_a4.pdf ..
-mv tutorial_mobile.pdf ..
-
